Win32: Provide make_install.sh with some vars, make it more robust.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@16930 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ packaging/win32/install.sh
@@ -1012,7 +1012,7 @@ function make_install() {
         # correct the 'dlname' in the libtool archives. We do not use these
         # files to dlopen the modules, so actually this is unneeded.
         # Also, in all installed .la files, remove the dependency_libs line
-        mv bin/*.dll gnucash || true
+        mv bin/*.dll gnucash 2>/dev/null || true
         for A in gnucash/*.la; do
             sed '/dependency_libs/d;s#../bin/##' $A > tmp ; mv tmp $A
         done
@@ -1022,7 +1022,7 @@ function make_install() {
 
         # gettext 0.17 installs translations to \share\locale, but not all
         # gnome packages have been recompiled against it
-        cp -a locale ../share && rm -rf locale
+        [ -d locale ] && cp -a locale ../share && rm -rf locale
     qpopd
 
     qpushd $_INSTALL_UDIR/etc/gconf/schemas
